More general renaming for toolbars
